refactor(native): extract console_keyboard from keyboard (#2831)
Move the per-console press dispatchers (handle_gameboy/gba/snes_key_pressed,
handle_single_joypad_key_pressed, handle_unmodified_key) into
keyboard/console_keyboard.rs; they delegate to super::hotkeys and
super::controller_mapping. mod.rs now holds only the entry points and
KeyOutcome. No behavior change.
